Replace pendingWrites count in BacklogFiller
BacklogFiller was using a pendingWrites count to determine when it
should return false from write or emit a `drain` event. Replacing this
is a queue of events that need to be added to the backlog; they are
added via a method `flush`, which is called from `write`. Write now
always returns false, and emits `drain` after the entire queue has been
written to Redis.