Fitler returns true if the user is in the datastore at all.
Previously, we were checking in a time range to make sure a user was
valid. This was causing some issues with logic and, as it turns out, not
even necessary. Now, the filter just checks to see if the user is in the
valid user datastore at all; if they are, they count as valid.